feat!: contract registers chains given in terms
 - move chainDetails from test to proposal
 - considered a constraint that terms includes agoric chainInfo
   but decide against to allow contract to fall back on agoricNames
